refactor the code, add PiniaStateManagement, create (CommentsList, CommentComponent) components
