+++ src/router/index.ts
@@ -3,7 +3,11 @@ import { createRouter, createWebHistory } from 'vue-router'
 const router = createRouter({
 history: createWebHistory(import.meta.env.BASE_URL),
 routes: [
-
+	{
+		path: '/',
+		name: 'Home',
+		component: () => import('@/views/HomeView.vue')
+	},
 ]
 })
