--- src/stores/counter.ts
@@ -1,6 +1,0 @@
-import { ref, computed } from 'vue'
-import { defineStore } from 'pinia'
-
-export const useCounterStore = defineStore('counter', () => {
-return {  }
-})
